+++ providers/denominator-route53/src/test/java/denominator/route53/Route53RoundRobinLiveTest.java
@@ -1,13 +1,7 @@
 package denominator.route53;
 
-import static com.google.common.base.Predicates.equalTo;
-import static com.google.common.base.Predicates.not;
-import static com.google.common.collect.Iterables.filter;
-
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-
-import com.google.common.collect.ImmutableList;
 
 import denominator.BaseRoundRobinLiveTest;
 
@@ -18,6 +12,5 @@ public class Route53RoundRobinLiveTest extends BaseRoundRobinLiveTest {
         Route53Connection connection = new Route53Connection();
         manager = connection.manager;
         setMutableZoneIfPresent(connection.mutableZone);
-        supportedRecordTypes = ImmutableList.copyOf(filter(supportedRecordTypes, not(equalTo("SSHFP"))));
     }
 }